# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Application {
  id: number;
  company: string;
  role: string;
  /** @nullable */
  location?: string | null;
  /**
   * Remote, Hybrid, or Onsite
   * @nullable
   */
  workType?: string | null;
  /** Wishlist, Applied, Interviewing, Offer, or Rejected */
  status: string;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /**
   * AI-computed match score 0-100
   * @nullable
   */
  matchScore?: number | null;
  /** @nullable */
  jobUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  appliedDate?: string | null;
  /**
   * Scheduled interview datetime (set when Applied)
   * @nullable
   */
  interviewDate?: string | null;
  /**
   * Offer acceptance deadline (set when Interviewing)
   * @nullable
   */
  offerDeadline?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ApplicationInput {
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  role: string;
  location?: string;
  workType?: string;
  status?: string;
  salaryMin?: number;
  salaryMax?: number;
  matchScore?: number;
  jobUrl?: string;
  notes?: string;
  appliedDate?: string;
  interviewDate?: string;
  offerDeadline?: string;
}

export interface ApplicationPatch {
  company?: string;
  role?: string;
  location?: string;
  workType?: string;
  status?: string;
  salaryMin?: number;
  salaryMax?: number;
  matchScore?: number;
  jobUrl?: string;
  notes?: string;
  appliedDate?: string;
  interviewDate?: string;
  offerDeadline?: string;
}

export interface TimelineEntry {
  id: number;
  applicationId: number;
  status: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface VsLastMonth {
  /** Percent change vs last month */
  applications: number;
  interviews: number;
  offers: number;
  rejections: number;
}

export interface DashboardStats {
  totalApplications: number;
  interviews: number;
  offers: number;
  rejections: number;
  avgResponseTimeDays: number;
  /** @nullable */
  avgMatchScore: number | null;
  vsLastMonth: VsLastMonth;
}

export interface FunnelEntry {
  stage: string;
  count: number;
  /** @nullable */
  conversionRate?: number | null;
}

export interface TimelineDataPoint {
  date: string;
  count: number;
}

export interface RoleCount {
  role: string;
  count: number;
  percentage: number;
}

export interface ResumeAnalysisInput {
  /** @minLength 1 */
  resumeText: string;
  /** @minLength 1 */
  jobDescription: string;
}

export interface ResumeBreakdown {
  impact: number;
  skills: number;
  experience: number;
  keywords: number;
  formatting: number;
}

export interface ResumeAnalysisResult {
  /** Overall ATS score 0-100 */
  atsScore: number;
  missingKeywords: string[];
  suggestions: string[];
  breakdown: ResumeBreakdown;
}

export interface InterviewQuestionsInput {
  /** @minLength 1 */
  jobType: string;
  /** Junior, Mid, Senior */
  experienceLevel?: string;
}

export interface InterviewQuestionsResult {
  jobType: string;
  behavioral: string[];
  technical: string[];
  systemDesign: string[];
}

export type ListApplicationsParams = {
  status?: string;
};