# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  role: string;
  location?: string;
  workType?: string;
  status?: string;
  salaryMin?: number;
  salaryMax?: number;
  matchScore?: number;
  jobUrl?: string;
  notes?: string;
  appliedDate?: Date;
  interviewDate?: Date;
  offerDeadline?: Date;
}