# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VsLastMonth } from "./vsLastMonth";

export interface DashboardStats {
  totalApplications: number;
  interviews: number;
  offers: number;
  rejections: number;
  avgResponseTimeDays: number;
  /** @nullable */
  avgMatchScore: number | null;
  vsLastMonth: VsLastMonth;
}