# lib/api-zod/src/generated/types/funnelEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FunnelEntry {
  stage: string;
  count: number;
  /** @nullable */
  conversionRate?: number | null;
}