# lib/api-zod/src/generated/types/interviewQuestionsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewQuestionsResult {
  jobType: string;
  behavioral: string[];
  technical: string[];
  systemDesign: string[];
}