# lib/api-zod/src/generated/types/resumeAnalysisInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeAnalysisInput {
  /** @minLength 1 */
  resumeText: string;
  /** @minLength 1 */
  jobDescription: string;
}