# lib/api-zod/src/generated/types/resumeAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResumeBreakdown } from "./resumeBreakdown";

export interface ResumeAnalysisResult {
  /** Overall ATS score 0-100 */
  atsScore: number;
  missingKeywords: string[];
  suggestions: string[];
  breakdown: ResumeBreakdown;
}