# lib/api-zod/src/generated/types/resumeBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeBreakdown {
  impact: number;
  skills: number;
  experience: number;
  keywords: number;
  formatting: number;
}