# lib/api-zod/src/generated/types/timelineEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEntry {
  id: number;
  applicationId: number;
  status: string;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}