# lib/api-zod/src/generated/types/vsLastMonth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VsLastMonth {
  /** Percent change vs last month */
  applications: number;
  interviews: number;
  offers: number;
  rejections: number;
}